parse int to Eth value

+++ scripts/getValuesInETH.py
@@ -4,6 +4,6 @@ eth_in_wei = 1000000000000000000
 scale = 100000
 
 def getEthValueInScale(amount):
-    print(round(amount/eth_value, 2) * eth_in_wei / scale)
+    print(int(round(amount/eth_value, 2) * eth_in_wei / scale))
 
 getEthValueInScale(float(sys.argv[1]))
